requirements.txt
Django==4.2.10
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
python-dotenv==1.0.1
django-chartjs==2.3.0
django-rest-framework==0.1.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
whitenoise==6.6.0
gunicorn==21.2.0
pandas>=2.0.0
requests>=2.31.0 